TracMercurial for 0.12: more robust filelog strategy
This is a fixup for r10512 and co, after realizing that the list of
files we're using in the filelog strategy as seeds to discover what is
the most recent change B in a folder before a given starting point in
the history A are not necessarily always present in every revisions in
the [A,B] range: there could be a several branches interspersed, and
in some the files could have been deleted or moved.

Issue reported and investigated by mike_mp@zzzcomputing.com.  Thanks
for the report and the fix!