1.0dev: use more correct query for report pagination.
The initial approach to report pagination (#216) relied on using
a subquery. In fact, there was no guarantee that the order of the
rows would be kept from the subquery to the main query, and this
was visible in some cases for some SQL engines.

In addition, we also support user defined sorting even in the
presence of a grouping by columns. In case the automated query
manipulation fails, it's possible to specify where the sort column
and the limit/offset clause will appear (using `@SORT_COLUMN` and
`@LIMIT_OFFSET@`, respectively).


Closes #10540.