1.0.2dev: ... and a follow-up to r11689, as there were still 2 test failures.
The functional tests TestAdminMilestoneDue and
TestAdminMilestoneDetailDue would fail with an "Invalid Date" error
when the "en_US.mo" compiled catalog is missing.

In fact, we should always have 'en_US' in the list of available
translations, as it's the "source" language so it shouldn't matter if
we have the compiled catalog or not. Ensuring we always have it fixes
the tests.