1.1.2dev: Hide the retargeting elements in the form if there are no tickets associated with the milestone. Refs #11336.
A hint is shown in place of the retargeting elements, //There are no tickets associated with this milestone.// The retargeting elements are shown now when there are no milestones to retarget the ticket to (i.e. the only choice is the //None// milestone). This is to make it obvious that tickets are being retargeted to another milestone, which in this case is the empty milestone.