1.1.2dev: Don't append db, web server or git version info directly to `systeminfo`. Refs #11587.
The version is instead provided through the `ISystemInfoProvider` interface.

The `GitConnector` and `SubversionConnector` are required components when a repository of their type is configured. When required, the Component cannot be disabled from the plugin admin page. With this change the behavior of the repository connectors matches that of database connectors.