1.3.2dev: decrease font-size globally (.9em)
The #633 fix introduced a font-size of 1em for the <body> and relative
sizes (em or %) everywhere else. This had the effect to make
everything look 25% bigger (13px -> 16px).  Maybe that was a bit too
extreme. Using .9em instead still increases the size of the elements
in the interface (base size 13px -> 14.4px), but only by a factor of
10%.

Also:
 - fix visual glitch when hovering on the mainnav tabs
 - made the ticket box a bit narrower (754px -> 806px instead of 835px)
 - made the wiki narrow mode a bit narrower (24px + 763px + 16px = 803px)